Delete user routes
Working on connecting the controller function to the userRouter
and the adminRouter, meaning I had to modify the the delete
controller functions to properly catch errors

+++ controllers/delete.js
@@ -2,24 +2,38 @@ import {
   deleteAllUsersExceptAdmin,
   deleteUserByID,
 } from "../prisma_queries/delete.js";
+import { findUserByID } from "../prisma_queries/find.js";
 
-export async function deleteUserSelf(req, res, next) {
+export async function removeUserSelf(req, res, next) {
   try {
     if (req.user.role == "ADMIN") {
-      return res.sentStatus(400);
+      return res.sendStatus(403);
     }
     await deleteUserByID(req.user.id);
-    res.sentStatus(200);
+    res.sendStatus(200);
   } catch (err) {
     return next(err);
   }
 }
 
-export async function deleteUserByAdmin(req, res, next) {
+export async function removeUserByAdmin(req, res, next) {
   try {
     const userId = Number(req.params.id);
+    const user = await findUserByID(userId);
+    if (user.role == "ADMIN") {
+      return res.sendStatus(403);
+    }
     await deleteUserByID(userId);
-    res.sentStatus(200);
+    res.sendStatus(200);
+  } catch (err) {
+    return next(err);
+  }
+}
+
+export async function removeAllUsersByAdmin(req, res, next) {
+  try {
+    await deleteAllUsersExceptAdmin();
+    res.sendStatus(200);
   } catch (err) {
     return next(err);
   }
